Move history persistence to history.go
